+++ lib/src/directive/models/resize-handle-type.ts
@@ -1,0 +1,13 @@
+// tslint:disable-next-line:interface-name
+export interface ResizeHandleStyle {
+  'n'?: string;
+  's'?: string;
+  'e'?: string;
+  'w'?: string;
+  'ne'?: string;
+  'nw'?: string;
+  'se'?: string;
+  'sw'?: string;
+}
+
+export type ResizeHandleType = string | ResizeHandleStyle;