+++ lib/src/public_api.ts
@@ -1,1 +1,4 @@
+export * from './directive/angular-draggable.directive';
+export * from './directive/angular-resizable.directive';
 export * from './angular-draggable.module';
+export * from './directive/models/position';